pyproject.toml
[project]
name = "python-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "openai-agents>=0.2.10",
    "python-dotenv>=1.1.1",
]
